feat: display CC recipients in message view for casedata and supportmanagement
Update API versions and data contracts to support ccRecipients.
Extract DTOs from controllers into separate files.
Show CC line in rendered messages for EMAIL type when CC recipients exist.
