+++ backend/src/config/api-config.ts
@@ -30,7 +30,7 @@ export const APIS = [
   },
   {
     name: 'case-data',
-    version: '11.14',
+    version: '12.1',
   },
   {
     name: 'supportmanagement',
