+++ backend/src/controllers/message.controller.ts
@@ -15,176 +15,20 @@ import {
 import { getOwnerStakeholder, getOwnerStakeholderEmail } from '@services/stakeholder.service';
 import { fileUploadOptions } from '@utils/fileUploadOptions';
 import { validateRequestBody } from '@utils/validate';
-import { IsArray, IsOptional, IsString } from 'class-validator';
 import { Body, Controller, Get, HttpCode, Param, Post, Put, Req, Res, UploadedFiles, UseBefore } from 'routing-controllers';
 import { OpenAPI, ResponseSchema } from 'routing-controllers-openapi';
 import { v4 as uuidv4 } from 'uuid';
 
 import { apiServiceName } from '@/config/api-config';
-import {
-  AttachmentResponse,
-  Classification,
-  EmailHeader,
-  Errand as ErrandDTO,
-  MessageResponse as IMessageResponse,
-  MessageResponseDirectionEnum,
-} from '@/data-contracts/case-data/data-contracts';
+import { Errand as ErrandDTO, MessageResponse as IMessageResponse } from '@/data-contracts/case-data/data-contracts';
 import { EmailAttachment, EmailRequest, SmsRequest, WebMessageAttachment, WebMessageRequest } from '@/data-contracts/messaging/data-contracts';
+import { AgnosticMessageResponse, DecisionMessageDto, MessageClassification, MessageDto, MessageResponse, SmsDto } from '@/dtos/message.dto';
 import { HttpException } from '@/exceptions/HttpException';
 import { isPT } from '@/services/application.service';
 import { logger } from '@/utils/logger';
 import { apiURL, base64Encode } from '@/utils/util';
 
-export enum MessageClassification {
-  'Efterfrågan komplettering' = 'COMPLETION_REQUEST',
-  'Informationsmeddelande' = 'INFORMATION',
-  'Hämta yttrande' = 'OBTAIN_OPINION',
-  'Intern dialog' = 'INTERNAL_COMMUNICATION',
-  // 'Övrigt' = 'OTHER
-}
-
-class MessageDto {
-  @IsString()
-  @IsOptional()
-  email!: string;
-  @IsString()
-  @IsOptional()
-  contactMeans!: string;
-  @IsString()
-  @IsOptional()
-  subject!: string;
-  @IsString()
-  text!: string;
-  @IsString()
-  attachUtredning!: string;
-  @IsString()
-  errandId!: string;
-  @IsString()
-  municipalityId!: string;
-  @IsString()
-  messageClassification!: MessageClassification;
-  @IsString()
-  reply_to!: string;
-  @IsString()
-  references!: string;
-  @IsOptional()
-  files!: Express.Multer.File[];
-}
-
-class SmsDto {
-  @IsString()
-  phonenumber!: string;
-  @IsString()
-  text!: string;
-  @IsString()
-  errandId!: string;
-  @IsString()
-  municipalityId!: string;
-}
-
-class DecisionMessageDto {
-  @IsString()
-  errandId!: string;
-}
-
-class MessageResponse implements IMessageResponse {
-  @IsOptional()
-  @IsString()
-  messageId?: string;
-  @IsOptional()
-  @IsString()
-  errandId?: number;
-  @IsOptional()
-  @IsString()
-  municipalityId?: string;
-  @IsOptional()
-  @IsString()
-  namespace?: string;
-  @IsOptional()
-  @IsString()
-  direction?: MessageResponseDirectionEnum;
-  @IsOptional()
-  @IsString()
-  familyId?: string;
-  @IsOptional()
-  @IsString()
-  externalCaseId?: string;
-  @IsOptional()
-  @IsString()
-  message?: string;
-  @IsOptional()
-  @IsString()
-  sent?: string;
-  @IsOptional()
-  @IsString()
-  subject?: string;
-  @IsOptional()
-  @IsString()
-  username?: string;
-  @IsOptional()
-  @IsString()
-  firstName?: string;
-  @IsOptional()
-  @IsString()
-  lastName?: string;
-  @IsOptional()
-  @IsString()
-  messageType?: string;
-  @IsOptional()
-  @IsString()
-  mobileNumber?: string;
-  @IsOptional()
-  @IsArray()
-  recipients?: string[];
-  @IsOptional()
-  @IsString()
-  email?: string;
-  @IsOptional()
-  @IsString()
-  htmlMessage?: string;
-  @IsOptional()
-  @IsString()
-  userId?: string;
-  @IsOptional()
-  @IsString()
-  viewed?: boolean;
-  @IsOptional()
-  @IsString()
-  classification?: Classification;
-  @IsOptional()
-  @IsArray()
-  attachments?: AttachmentResponse[];
-  @IsOptional()
-  @IsArray()
-  emailHeaders?: EmailHeader[];
-}
-
-export interface AgnosticMessageResponse {
-  messageId: string;
-}
-
-export interface WebMessageResponse {
-  messageId: string;
-  deliveries: {
-    deliveryId: string;
-    messageType: string;
-    status: string;
-  }[];
-}
-
-export interface LetterResponse {
-  batchId: string;
-  messages: [
-    {
-      messageId: string;
-      deliveries: {
-        deliveryId: string;
-        messageType: 'DIGITAL_MAIL' | 'SNAIL_MAIL';
-        status: string;
-      }[];
-    },
-  ];
-}
+export { AgnosticMessageResponse, LetterResponse, MessageClassification, WebMessageResponse } from '@/dtos/message.dto';
 
 const MESSAGE_SUBJECT = isPT() ? 'Meddelande gällande er ansökan om parkeringstillstånd' : 'Meddelande från MEX';
