+++ backend/src/data-contracts/billing-data-collector/data-contracts.ts
@@ -120,18 +120,18 @@ export interface PageableObject {
   /** @format int64 */
   offset?: number;
   unpaged?: boolean;
-  sort?: SortObject;
   paged?: boolean;
   /** @format int32 */
   pageNumber?: number;
   /** @format int32 */
   pageSize?: number;
+  sort?: SortObject;
 }
 
 export interface SortObject {
   empty?: boolean;
-  unsorted?: boolean;
   sorted?: boolean;
+  unsorted?: boolean;
 }
 
 /**
