+++ backend/src/data-contracts/billingpreprocessor/data-contracts.ts
@@ -228,10 +228,10 @@ export interface InvoiceFileStatus {
 }
 
 export interface PageBillingRecord {
-  /** @format int64 */
-  totalElements?: number;
   /** @format int32 */
   totalPages?: number;
+  /** @format int64 */
+  totalElements?: number;
   /** @format int32 */
   size?: number;
   content?: BillingRecord[];
@@ -249,19 +249,19 @@ export interface PageBillingRecord {
 export interface PageableObject {
   /** @format int64 */
   offset?: number;
+  unpaged?: boolean;
   sort?: SortObject;
   paged?: boolean;
   /** @format int32 */
   pageNumber?: number;
   /** @format int32 */
   pageSize?: number;
-  unpaged?: boolean;
 }
 
 export interface SortObject {
   empty?: boolean;
-  sorted?: boolean;
   unsorted?: boolean;
+  sorted?: boolean;
 }
 
 export enum GetFileStatusesForMonthParamsMonthEnum {
