+++ frontend/src/casedata/components/errand/tabs/messages/render-message-reciever.component.tsx
@@ -42,7 +42,7 @@ interface EmailRecipientsProps {
   recipients: string[];
 }
 
-const EmailRecipients: FC<EmailRecipientsProps> = ({ recipients }) => {
+export const EmailRecipients: FC<EmailRecipientsProps> = ({ recipients }) => {
   const [isExpanded, setIsExpanded] = useState(false);
 
   const totalCount = recipients.length;
