+++ frontend/src/data-contracts/backend/data-contracts.ts
@@ -156,27 +156,27 @@ export interface CPageBillingRecord {
 }
 
 export interface ExtraParametersDto {
-  "application.reason"?: string;
-  "application.role"?: ExtraParametersDtoApplicationRoleEnum;
-  "application.applicant.capacity"?: string;
-  "application.applicant.testimonial"?: ExtraParametersDtoApplicationApplicantTestimonialEnum;
-  "application.applicant.signingAbility"?: ExtraParametersDtoApplicationApplicantSigningAbilityEnum;
-  "disability.aid"?: string;
-  "disability.walkingAbility"?: ExtraParametersDtoDisabilityWalkingAbilityEnum;
-  "disability.walkingDistance.beforeRest"?: string;
-  "disability.walkingDistance.max"?: string;
-  "disability.duration"?: string;
-  "disability.canBeAloneWhileParking"?: ExtraParametersDtoDisabilityCanBeAloneWhileParkingEnum;
-  "disability.canBeAloneWhileParking.note"?: string;
-  "consent.contact.doctor"?: ExtraParametersDtoConsentContactDoctorEnum;
-  "consent.view.transportationServiceDetails"?: ExtraParametersDtoConsentViewTransportationServiceDetailsEnum;
-  "application.lostPermit.policeReportNumber"?: string;
-  "application.renewal.changedCircumstances"?: ExtraParametersDtoApplicationRenewalChangedCircumstancesEnum;
-  "application.renewal.expirationDate"?: string;
-  "application.renewal.medicalConfirmationRequired"?: string;
-  "artefact.permit.number"?: string;
-  "artefact.permit.status"?: string;
-  "application.supplement.dueDate"?: string;
+  'application.reason'?: string;
+  'application.role'?: ExtraParametersDtoApplicationRoleEnum;
+  'application.applicant.capacity'?: string;
+  'application.applicant.testimonial'?: ExtraParametersDtoApplicationApplicantTestimonialEnum;
+  'application.applicant.signingAbility'?: ExtraParametersDtoApplicationApplicantSigningAbilityEnum;
+  'disability.aid'?: string;
+  'disability.walkingAbility'?: ExtraParametersDtoDisabilityWalkingAbilityEnum;
+  'disability.walkingDistance.beforeRest'?: string;
+  'disability.walkingDistance.max'?: string;
+  'disability.duration'?: string;
+  'disability.canBeAloneWhileParking'?: ExtraParametersDtoDisabilityCanBeAloneWhileParkingEnum;
+  'disability.canBeAloneWhileParking.note'?: string;
+  'consent.contact.doctor'?: ExtraParametersDtoConsentContactDoctorEnum;
+  'consent.view.transportationServiceDetails'?: ExtraParametersDtoConsentViewTransportationServiceDetailsEnum;
+  'application.lostPermit.policeReportNumber'?: string;
+  'application.renewal.changedCircumstances'?: ExtraParametersDtoApplicationRenewalChangedCircumstancesEnum;
+  'application.renewal.expirationDate'?: string;
+  'application.renewal.medicalConfirmationRequired'?: string;
+  'artefact.permit.number'?: string;
+  'artefact.permit.status'?: string;
+  'application.supplement.dueDate'?: string;
 }
 
 export interface Attachment {
@@ -406,6 +406,7 @@ export interface MessageResponse {
   messageType?: string;
   mobileNumber?: string;
   recipients?: any[];
+  ccRecipients?: string[];
   email?: string;
   htmlMessage?: string;
   userId?: string;
@@ -568,6 +569,7 @@ export interface CCommunication {
   target?: string;
   internal?: boolean;
   viewed?: boolean;
+  ccRecipients?: string[];
   emailHeaders?: string;
   communicationAttachments: CCommunicationAttachment[];
 }
@@ -612,65 +614,65 @@ export interface TemplateSelector {
 }
 
 export enum CBillingRecordTypeEnum {
-  EXTERNAL = "EXTERNAL",
-  INTERNAL = "INTERNAL",
+  EXTERNAL = 'EXTERNAL',
+  INTERNAL = 'INTERNAL',
 }
 
 export enum CBillingRecordStatusEnum {
-  NEW = "NEW",
-  APPROVED = "APPROVED",
-  INVOICED = "INVOICED",
-  REJECTED = "REJECTED",
+  NEW = 'NEW',
+  APPROVED = 'APPROVED',
+  INVOICED = 'INVOICED',
+  REJECTED = 'REJECTED',
 }
 
 export enum ExtraParametersDtoApplicationRoleEnum {
-  SELF = "SELF",
-  GUARDIAN = "GUARDIAN",
-  CUSTODIAN = "CUSTODIAN",
+  SELF = 'SELF',
+  GUARDIAN = 'GUARDIAN',
+  CUSTODIAN = 'CUSTODIAN',
 }
 
 export enum ExtraParametersDtoApplicationApplicantTestimonialEnum {
-  True = "true",
-  False = "false",
+  True = 'true',
+  False = 'false',
 }
 
 export enum ExtraParametersDtoApplicationApplicantSigningAbilityEnum {
-  True = "true",
-  False = "false",
+  True = 'true',
+  False = 'false',
 }
 
 export enum ExtraParametersDtoDisabilityWalkingAbilityEnum {
-  True = "true",
-  False = "false",
+  True = 'true',
+  False = 'false',
 }
 
 export enum ExtraParametersDtoDisabilityCanBeAloneWhileParkingEnum {
-  True = "true",
-  False = "false",
+  True = 'true',
+  False = 'false',
 }
 
 export enum ExtraParametersDtoConsentContactDoctorEnum {
-  True = "true",
-  False = "false",
+  True = 'true',
+  False = 'false',
 }
 
 export enum ExtraParametersDtoConsentViewTransportationServiceDetailsEnum {
-  True = "true",
-  False = "false",
+  True = 'true',
+  False = 'false',
 }
 
 export enum ExtraParametersDtoApplicationRenewalChangedCircumstancesEnum {
-  Y = "Y",
-  N = "N",
+  Y = 'Y',
+  N = 'N',
 }
 
 export enum CCommunicationDirectionEnum {
-  INBOUND = "INBOUND",
-  OUTBOUND = "OUTBOUND",
+  INBOUND = 'INBOUND',
+  OUTBOUND = 'OUTBOUND',
 }
 
 export enum CCommunicationCommunicationTypeEnum {
-  SMS = "SMS",
-  EMAIL = "EMAIL",
-  WEB_MESSAGE = "WEB_MESSAGE",
+  SMS = 'SMS',
+  EMAIL = 'EMAIL',
+  WEB_MESSAGE = 'WEB_MESSAGE',
 }
