+++ frontend/src/supportmanagement/services/support-message-service.ts
@@ -45,6 +45,7 @@ export interface Message {
   conversationId?: string;
   messageId?: string;
   recipients: string[];
+  ccRecipients: string[];
 }
 
 const getClosingMessageBody = (): string => {
@@ -351,6 +352,7 @@ export const buildTree = (_list: Message[]) => {
           },
           children: [],
           recipients: [],
+          ccRecipients: [],
         };
         dummyParent?.children?.push(nodesMap.get(id)!);
         roots.push(dummyParent);
